Rename: 마이페이지 유저 예매내역 페이지 이름 변경 (#45)
